backend/requirements.txt
fastapi==0.89.1
uvicorn==0.22.0
sqlalchemy==2.0.12
pydantic==1.10.7
python-multipart==0.0.6
bcrypt==4.0.1
pyjwt==2.6.0
numpy>=1.26.0
google-generativeai>=0.3.0
aiohttp>=3.9.0
pydantic-settings==0.2.5